[#166] refactor: 광고 추가 및 파일 삭제
