+++ src/app/(route)/(main)/market/enchant/page.tsx
@@ -1,3 +1,4 @@
+import SideAd from '@/app/_components/adsense/SideAd';
 import { insertUpgradeType } from '@/app/_components/enchant/utils/insertUpgradeType';
 import BasicContainer from '@/app/_components/layout/BasicContainer';
 import Column from '@/app/_components/layout/Column';
@@ -19,18 +20,22 @@ const Page = () => {
   ];
 
   return (
-    <BasicContainer className="p-2 lg:mt-10">
-      <div className="mx-auto flex w-full max-w-md flex-col flex-wrap justify-center gap-3 rounded-md border-borderColor/50 p-2 md:max-w-5xl md:flex-row md:p-6 lg:p-10">
-        <Column className="z-10 max-w-md">
-          <EnchantRankTable enchantData={allEnchantList} />
-          <EnchantTableInputFilter />
-        </Column>
-        <EnchantDropAndDetail />
-      </div>
-      <div className="my-4 hidden text-center text-xs sm:block">
-        거래량이 적은 인챈트는 가격 정보가 표시되지 않습니다
-      </div>
-    </BasicContainer>
+    <>
+      <SideAd dataSlot="2056348937" position="left" />
+      <BasicContainer className="p-2 lg:mt-10">
+        <div className="mx-auto flex w-full max-w-md flex-col flex-wrap justify-center gap-3 rounded-md border-borderColor/50 p-2 md:max-w-5xl md:flex-row md:p-6 lg:p-10">
+          <Column className="z-10 max-w-md">
+            <EnchantRankTable enchantData={allEnchantList} />
+            <EnchantTableInputFilter />
+          </Column>
+          <EnchantDropAndDetail />
+        </div>
+        <div className="my-4 hidden text-center text-xs sm:block">
+          거래량이 적은 인챈트는 가격 정보가 표시되지 않습니다
+        </div>
+      </BasicContainer>
+      <SideAd dataSlot="1601053361" position="right" />
+    </>
   );
 };
