+++ src/app/(route)/(main)/raid/page.tsx
@@ -1,12 +1,17 @@
+import SideAd from '@/app/_components/adsense/SideAd';
 import BasicContainer from '@/app/_components/layout/BasicContainer';
 import { LimitTable, LimitTableMenuBar } from '@/app/_features/raid';
 
 const Page = () => {
   return (
-    <BasicContainer className="flex-1 gap-2">
-      <LimitTableMenuBar />
-      <LimitTable />
-    </BasicContainer>
+    <>
+      <SideAd dataSlot="2056348937" position="left" />
+      <BasicContainer className="flex-1 gap-2">
+        <LimitTableMenuBar />
+        <LimitTable />
+      </BasicContainer>
+      <SideAd dataSlot="1601053361" position="right" />
+    </>
   );
 };
 export default Page;
