+++ src/app/(route)/(main)/raidinfo/page.tsx
@@ -1,10 +1,15 @@
+import SideAd from '@/app/_components/adsense/SideAd';
 import { RaidInfoTable } from '@/app/_features/raidinfo';
 
 const Page = () => {
   return (
-    <div className="px-4">
-      <RaidInfoTable />
-    </div>
+    <>
+      <SideAd dataSlot="2056348937" position="left" />
+      <div className="px-4">
+        <RaidInfoTable />
+      </div>
+      <SideAd dataSlot="1601053361" position="right" />
+    </>
   );
 };
