+++ src/app/_features/gold/types.ts
@@ -6,14 +6,14 @@ export type MarketRankList<T> = T extends 'buy'
   ? MarKetBuyRankList
   : MarKetSellRankList;
 
-export interface MarKetBuyRankList {
+interface MarKetBuyRankList {
   buy_gold: {
     cairde_name: string;
     buy_gold: number;
   }[];
 }
 
-export interface MarKetSellRankList {
+interface MarKetSellRankList {
   sell_gold: {
     cairde_name: string;
     sell_gold: number;
