--- src/app/_features/preview/components/preview/infusion/InfusionGroup.tsx
@@ -1,22 +1,0 @@
-import { InfusionGroupProps } from '@/app/_type/infusionType';
-import InfusionGrid from './InfusionGrid';
-
-const InfusionGroup = ({
-  title,
-  infusionList,
-  selectedHandler,
-  selectedValue,
-}: InfusionGroupProps) => {
-  return (
-    <li className="my-4">
-      <h3 className="text-base text-white">{title}</h3>
-      <InfusionGrid
-        infusionList={infusionList}
-        selectedHandler={selectedHandler}
-        selectedValue={selectedValue}
-      />
-    </li>
-  );
-};
-
-export default InfusionGroup;
