--- src/app/_features/preview/components/preview/infusion/InfusionValues.tsx
@@ -1,20 +1,0 @@
-import Column from '@/app/_components/layout/Column';
-import { InfusionValuesProps } from '@/app/_type/infusionType';
-
-const InfusionValues = ({ infusionValue }: InfusionValuesProps) => {
-  return (
-    <Column className="min-h-24 flex-1 rounded-md border border-borderColor bg-black/20 p-1 font-sans text-xs">
-      {infusionValue.map((infusion) => (
-        <div
-          className="flex items-center gap-1 px-1"
-          key={infusion.stat_name + infusion.stat_value}
-        >
-          <p className="flex justify-start">• {infusion.stat_name}</p>
-          <p className="flex flex-1 items-center">{infusion.stat_value}</p>
-        </div>
-      ))}
-    </Column>
-  );
-};
-
-export default InfusionValues;
